Refactor the utils.ts file and Utils directory
I divided the responsibilities between the files. I renamed the files.
