+++ frontend/src/utils/statsUtils.ts
