+++ frontend/src/utils/textUtils.ts
